+++ src/index.scss
@@ -28,7 +28,6 @@
 @import "./components/button-more/button-more";
 @import "./components/table-transactions/table-transactions";
 @import "./components/chart/chart";
-@import "./components/accounts/accounts-component/accounts-component";
+@import "components/accounts/accounts/accounts";
 @import "./components/custom-select/custom-select";
 @import "./components/about-dialog/about-dialog";
-
